+++ R/data_help_files.R
@@ -5,3 +5,11 @@
 #' @docType data
 #' @format A data frame
 NULL
+
+#' @rdname past_results_v2
+#' @name past_results_v2
+#' @title Benchmarking results
+#' @description A summary of past benchmarks. 
+#' @docType data
+#' @format A data frame
+NULL