+++ R/select_results.R
@@ -1,69 +1,39 @@
 #' Selecting results
 #' 
-#' Selects and aggregates over the \code{past_results} data set or the 
+#' Selects and aggregates over the \code{past_results_v2} data set or the 
 #' \code{results} input data set..
 #' @return A data frame
-#' @param parallel Default \code{FALSE}. Should the parallel benchmarks be returned
+#' @param results Default \code{NULL}. If \code{NULL} the \code{past_results_v2}
+#' data set is used. Otherwise, the input data set.
 #' @inheritParams plot_past
 #' @export
 #' @examples 
-#' select_results("prog", byte_optimize=TRUE)
-#' select_results("matrix_fun", blas_optimize=TRUE)
+#' select_results("prog", blas_optimize = NULL)
 select_results = function(test_group, 
                           results = NULL,
-                          byte_optimize = NULL, 
                           blas_optimize = NULL, 
-                          parallel = FALSE) {
+                          cores = 0) {
+  
+  if (is.null(blas_optimize)) blas_optimize = c(FALSE, TRUE)
+  
   ## Load past data
   tmp_env = new.env()
-  data(past_results, package="benchmarkmeData", envir = tmp_env)
-  if(!is.null(results)) {
-  
-    results = results %>%
-      group_by(test_group) %>%
-      mutate(std_time = time/if(all(cores==0)) min(time) else mean(time[cores == 1]))
-  }
-  
-  
-  results = rbind(tmp_env$past_results, results)
+  data(past_results_v2, package = "benchmarkmeData", envir = tmp_env)
 
-  if(!is.null(byte_optimize)) {
-    if(byte_optimize) {
-      results = results[results$byte_optimize > 0.5,]
-    } else {
-      results = results[results$byte_optimize < 0.5,]
-    }
-  }
-  
-  if(!is.null(blas_optimize)) {
-    results = results[results$blas_optimize == blas_optimize,]
-  }
-  
-  if(length(test_group) > 1) test_group = test_group[1]
-  if(test_group %in% c("read", "write")) test_group = paste0(test_group, c(5, 50, 200))
-  
-  results = results[results$test_group %in% test_group,]
-  #results = results[results$cores == cores,]
-  
-  ## Aggregate over test
-  ## Ensure that we have timings for all required tests.
-  #results = aggregate(time ~ id + byte_optimize + cpu + date + 
-  #                      sysname + blas_optimize + test_group + ram + cores + std_time, 
-  #                    data=results, 
-  #                    FUN=sum)
-  
-  results = results[!is.na(results$time), ]
-  results$test_group = factor(results$test_group, levels=test_group)
-  results = results[order(results$time), ]
-  if(parallel) {
-    results = results[results$cores != 0,]
-    results$time = results$std_time
-  } else {
-    results = results[results$cores == 0,]
-  }
-  
-  
-  results$rank = 1:nrow(results)
-  
+  # Format data
+  results = results %>%
+    bind_rows(tmp_env$past_results_v2) %>%
+    filter(blas_optimize %in% !!blas_optimize) %>% 
+    filter(test_group == !!test_group) %>%
+    filter(cores %in% !!cores) %>%
+    filter(!is.na(time)) %>%
+    group_by(id, cpu, date, sysname, blas_optimize, test_group, ram, cores) %>%
+    summarise(time = sum(time)) %>%
+    ungroup()
+
+  results = results %>%
+    arrange(time) %>%
+    mutate(rank = 1:length(time)) %>%
+    select(rank, everything())
   results
 }